requirements.txt
requests
pydantic
pvporcupine
pyaudio
openai-whisper
sounddevice
